# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sales Performance Index API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /specialists:
    get:
      operationId: listSpecialists
      tags: [specialists]
      summary: List active specialists
      parameters:
        - name: archived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of specialists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Specialist"
    post:
      operationId: createSpecialist
      tags: [specialists]
      summary: Create a specialist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialist"
  /specialists/{id}:
    get:
      operationId: getSpecialist
      tags: [specialists]
      summary: Get a specialist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Specialist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialist"
        "404":
          description: Not found
    patch:
      operationId: updateSpecialist
      tags: [specialists]
      summary: Update a specialist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialistUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialist"
    delete:
      operationId: deleteSpecialist
      tags: [specialists]
      summary: Delete a specialist permanently
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /specialists/{id}/archive:
    patch:
      operationId: archiveSpecialist
      tags: [specialists]
      summary: Archive or unarchive a specialist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialist"
  /specialists/{id}/stats:
    get:
      operationId: getSpecialistStats
      tags: [specialists]
      summary: Get performance stats for a specialist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Specialist stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecialistStats"

  # Evaluations
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: specialistId
          in: query
          schema:
            type: integer
        - name: evaluationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task for a specialist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard
...
components:
  schemas:
...
    Specialist:
      type: object
      required: [id, firstName, lastName, position, department, hireDate, status, archived, spiScore, evaluationCount]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        department:
          type: string
        hireDate:
          type: string
        manager:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        archived:
          type: boolean
        spiScore:
          type: ["number", "null"]
        evaluationCount:
          type: integer
        monthlyTarget:
          type: ["integer", "null"]
        createdAt:
          type: string
    SpecialistInput:
      type: object
      required: [firstName, lastName, position, department, hireDate, status]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        department:
          type: string
        hireDate:
          type: string
        manager:
          type: string
        status:
          type: string
          enum: [active, inactive]
        monthlyTarget:
          type: integer
    SpecialistUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        department:
          type: string
        hireDate:
          type: string
        manager:
          type: string
        status:
          type: string
          enum: [active, inactive]
        monthlyTarget:
          type: ["integer", "null"]
    ArchiveInput:
      type: object
      required: [archived]
      properties:
        archived:
          type: boolean
    SpecialistStats:
      type: object
      required: [specialistId, averageScore, evaluationCount, bestSection, worstSection, monthlyTrend, sectionScores, radarData]
      properties:
        specialistId:
          type: integer
        averageScore:
          type: ["number", "null"]
        evaluationCount:
          type: integer
        bestSection:
          type: ["string", "null"]
        worstSection:
          type: ["string", "null"]
        monthlyTrend:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyTrendPoint"
        sectionScores:
          type: array
          items:
            $ref: "#/components/schemas/SectionScore"
        radarData:
          type: array
          items:
            $ref: "#/components/schemas/RadarPoint"

    # Evaluations
...
    Task:
      type: object
      required: [id, specialistId, description, deadline, createdBy, createdAt]
      properties:
        id:
          type: integer
        evaluationId:
          type: ["integer", "null"]
        specialistId:
          type: integer
        description:
          type: string
        deadline:
          type: string
        createdBy:
          type: integer
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [specialistId, description, deadline]
      properties:
        evaluationId:
          type: integer
        specialistId:
          type: integer
        description:
          type: string
        deadline:
          type: string

    # Dashboard
...
    MonthlyTrendPoint:
      type: object
      required: [month, averageScore, count]
      properties:
        month:
          type: string
        averageScore:
          type: ["number", "null"]
        count:
          type: integer
    SectionScore:
      type: object
      required: [sectionId, sectionName, averageScore]
      properties:
        sectionId:
          type: integer
        sectionName:
          type: string
        averageScore:
          type: ["number", "null"]
    RadarPoint:
      type: object
      required: [subject, score, fullMark]
      properties:
        subject:
          type: string
        score:
          type: number
        fullMark:
          type: number