# lib/api-zod/src/generated/types/specialist.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sales Performance Index API
 * OpenAPI spec version: 0.1.0
 */
import type { SpecialistStatus } from './specialistStatus';

export interface Specialist {
  id: number;
  firstName: string;
  lastName: string;
  position: string;
  department: string;
  hireDate: string;
  /** @nullable */
  manager?: string | null;
  status: SpecialistStatus;
  archived: boolean;
  /** @nullable */
  spiScore: number | null;
  evaluationCount: number;
  /** @nullable */
  monthlyTarget?: number | null;
  createdAt?: string;
}