# lib/api-zod/src/generated/types/specialistUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sales Performance Index API
 * OpenAPI spec version: 0.1.0
 */
import type { SpecialistUpdateStatus } from './specialistUpdateStatus';

export interface SpecialistUpdate {
  firstName?: string;
  lastName?: string;
  position?: string;
  department?: string;
  hireDate?: string;
  manager?: string;
  status?: SpecialistUpdateStatus;
  /** @nullable */
  monthlyTarget?: number | null;
}