# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sales Performance Index API
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  /** @nullable */
  evaluationId?: number | null;
  specialistId: number;
  description: string;
  deadline: string;
  createdBy: number;
  createdAt: string;
}