# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sales Performance Index API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskInput {
  evaluationId?: number;
  specialistId: number;
  description: string;
  deadline: string;
}